+++ src/shared/util.odin
@@ -1,11 +1,11 @@
 package shared
 
-util_hash :: proc(data : string) -> u32 {
+util_hash :: proc(data : string) -> u16 {
 	i, size := int(0), len(data)
-	hash := u32(5381);
+	hash := u16(5381);
 
 	for i < size {
-		hash = ((hash << 5) + hash) + u32((u8(data[i])))
+		hash = ((hash << 5) + hash) + u16((u8(data[i])))
 		i += 1
 	}
